+++ ChangeLog.md
@@ -27,6 +27,12 @@
   * Set-AzureRMBackupProtectionPolicy
   * Get-AzureRMBackupRecoveryPoint
   * Restore-AzureRMBackupItem
+ * SQL Server VM cmdlets (ARM)
+  * New-AzureVMSqlServerAutoPatchingConfig
+  * New-AzureVMSqlServerAutoBackupConfig
+  * Set-AzureVMSqlServerExtension
+  * Get-AzureVMSqlServerExtension
+  * Remove-AzureVMSqlServerExtension
 	
 ## 2015.08.17 version 0.9.7
 * Azure Profile cmdlets